Update set-defaults-for-enterprise-deployments.md
The installer enforces a 50 character limit for the cache path.  Since we are increasing discoverability of this group policy in an ADMX file available in Intune and Microsoft Download Center, it makes sense to update this document with this information so that admins don't add a path that is too long.